requirements.txt
wxpython
pyzmq
configobj
pypubsub
python-vlc
cx_Freeze
